Created skeleton for solution.cpp in lists/find_the_loop
This includes test cases and a simple main function to run them.
Comments are included about what the test cases are testing and
generally what each expects. All that’s left to be done is to fill in
the method!

In addition I moved some code in Linked List common files around to
allow for compilation of templated methods. Includes minor changes
elsewhere.